% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemMatrixConversion.R
\name{as.item.matrix}
\alias{as.item.matrix}
\title{Function for converting qcba rules dataframe structure to arules
itemMatrix structure.}
\usage{
as.item.matrix(qcbaRuleModel, trainingData)
}
\arguments{
\item{qcbaRuleModel}{qcba rule model}

\item{trainingData}{data which was used for training the rule model}
}
\description{
Function for converting qcba rules dataframe structure to arules
itemMatrix structure.
}
\examples{
  library(arc)
  library(qCBA)

  data <- cars

  rmCBA <- cba(data, classAtt=colnames(data)[length(colnames(data))])
  rmqCBA <- qcba(cbaRuleModel=rmCBA,datadf=data)

  itemMatrixRules <- as.item.matrix(rmqCBA, data)

}
