% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemMatrixConversion.R
\name{as.qcba.rules}
\alias{as.qcba.rules}
\title{Function for converting arules rules data structure to qcba dataframe.}
\usage{
as.qcba.rules(rules)
}
\arguments{
\item{rules}{rules to be converted to qcba data structure}
}
\description{
Function for converting arules rules data structure to qcba dataframe.
}
\examples{
  library(arc)
  library(qCBA)

  data <- cars

  rmCBA <- cba(data, classAtt=colnames(data)[length(colnames(data))])
  rmqCBA <- qcba(cbaRuleModel=rmCBA,datadf=data)

  itemMatrixRules <- as.item.matrix(rmqCBA, data)
  qcbaRules <- as.qcba.rules(itemMatrixRules)


}
