% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalReader.R
\name{createIntervalReader}
\alias{createIntervalReader}
\title{Helper function for creating an intervalReader object.}
\usage{
createIntervalReader(numberSeparator = ";", closedBrackets = c("(",
  ")"), openBrackets = c("[", "]"), infinities = c("-Inf", "Inf"),
  bracketLen = 1, decimalSeparator = ".")
}
\arguments{
\item{numberSeparator}{string for separating numbers in an interval}

\item{closedBrackets}{vector of length 2 for parsing closed brackets}

\item{openBrackets}{vector of length 2 for parsing open brackets}

\item{infinities}{vector of length 2 for parsing positive and negative infinities}

\item{bracketLen}{how long in chars a bracket is}

\item{decimalSeparator}{string used as a decimal separator in an interval}
}
\value{
new intervalReader object
}
\description{
Helper function for creating an intervalReader object.
}
