% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\docType{methods}
\name{explainInstances,explanationObject,data.frame-method}
\alias{explainInstances,explanationObject,data.frame-method}
\title{A method for exlaining instances}
\usage{
\S4method{explainInstances}{explanationObject,data.frame}(theObject,
  dataToExplain)
}
\arguments{
\item{theObject}{an explanation object}

\item{dataToExplain}{data to be explained}
}
\description{
A method for exlaining instances
}
