% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\name{explainInstances}
\alias{explainInstances}
\title{A generic method for explaining instances}
\usage{
explainInstances(theObject, dataToExplain)
}
\arguments{
\item{theObject}{an explanation object}

\item{dataToExplain}{data to be explained}
}
\description{
A generic method for explaining instances
}
