% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanations.R
\name{explainPrediction.CBARuleModel}
\alias{explainPrediction.CBARuleModel}
\title{Function for investigating which rule classified an instance.}
\usage{
explainPrediction.CBARuleModel(object, data, discretize = TRUE, ...)
}
\arguments{
\item{object}{CBARuleModel object}

\item{data}{testing data for prediction}

\item{discretize}{determines if data should be discretized or not}

\item{...}{other arguments (currently not used)}
}
\value{
vector of rule IDs
}
\description{
Function for investigating which rule classified an instance.
}
