% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanations.R
\name{explainQCBA}
\alias{explainQCBA}
\title{Function for explaning QCBA.}
\usage{
explainQCBA(expl, rulesText, rules, rulesID)
}
\arguments{
\item{expl}{an explanation object}

\item{rulesText}{Text of the rules.}

\item{rules}{Rule dataframe.}

\item{rulesID}{IDs of the rules that classified an instance}
}
\value{
dataframe containing explanations for rules
}
\description{
Function for explaning QCBA.
}
