% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\docType{methods}
\name{explainRuleModel,explanationObject,data.frame-method}
\alias{explainRuleModel,explanationObject,data.frame-method}
\title{Method for explaining a rule model}
\usage{
\S4method{explainRuleModel}{explanationObject,data.frame}(theObject, data)
}
\arguments{
\item{theObject}{an explanation objectg}

\item{data}{data to be explained}
}
\description{
Method for explaining a rule model
}
