% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\name{explainRuleModel}
\alias{explainRuleModel}
\title{Generic method for explaining a rule model}
\usage{
explainRuleModel(theObject, data)
}
\arguments{
\item{theObject}{an explanation objectg}

\item{data}{data to be explained}
}
\description{
Generic method for explaining a rule model
}
