% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\docType{class}
\name{explanationObject-class}
\alias{explanationObject-class}
\alias{explanationObject}
\title{Explanation object}
\description{
Explanation object
}
\section{Slots}{

\describe{
\item{\code{ruleDataFrame}}{data frame of mined rules}

\item{\code{dataCount}}{number of training data}

\item{\code{intervalReader}}{an intervalReader object}

\item{\code{ruleModel}}{a cba rule model}
}}

