% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanations.R
\name{getExplanationsDataframe}
\alias{getExplanationsDataframe}
\title{Function for getting the explanation dataframe.}
\usage{
getExplanationsDataframe(expl, dataToExplain, ruleModel)
}
\arguments{
\item{expl}{an explanation object}

\item{dataToExplain}{data to be explained row by row}

\item{ruleModel}{model to use to explain}
}
\value{
explanation dataframe
}
\description{
Function for getting the explanation dataframe.
}
