% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanations.R
\name{getQCBAConflictingRuleText}
\alias{getQCBAConflictingRuleText}
\title{Function for getting information about the next best rule which has different
class from the current rule.}
\usage{
getQCBAConflictingRuleText(expl, ruleIndex)
}
\arguments{
\item{expl}{an explanation object}

\item{ruleIndex}{index of the rule we want to get conflicting text for}
}
\value{
conflicting rule text
}
\description{
Function for getting information about the next best rule which has different
class from the current rule.
}
