% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\docType{methods}
\name{initializeExplanation}
\alias{initializeExplanation}
\alias{initializeExplanation,explanationObject,CBARuleModel,data.frame,ANY-method}
\title{Generic method for initializing explanations}
\usage{
initializeExplanation(theObject, ruleModel, trainingData, ruleModelQCBA)


  \S4method{initializeExplanation}{explanationObject,CBARuleModel,data.frame,ANY}(theObject,
  ruleModel, trainingData)
}
\arguments{
\item{theObject}{an explanation objectg}

\item{ruleModel}{a rule model of CBA}

\item{trainingData}{training data of the model}

\item{ruleModelQCBA}{a rule model of QCBA}
}
\description{
Generic method for initializing explanations

Method for initializing explanations when the model is
 trained by cba
}
