% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanationObject.R
\docType{methods}
\name{initializeExplanation,explanationObject,CBARuleModel,data.frame,qCBARuleModel-method}
\alias{initializeExplanation,explanationObject,CBARuleModel,data.frame,qCBARuleModel-method}
\title{Method for initializing explanations when the model is trainined by qcba}
\usage{

  \S4method{initializeExplanation}{explanationObject,CBARuleModel,data.frame,qCBARuleModel}(theObject,
  ruleModel, trainingData, ruleModelQCBA)
}
\arguments{
\item{theObject}{an explanation objectg}

\item{ruleModel}{a rule model of CBA}

\item{trainingData}{training data of the model}

\item{ruleModelQCBA}{a rule model of QCBA}
}
\description{
Method for initializing explanations when the model is trainined by qcba
}
