% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalReader.R
\docType{class}
\name{intervalReader-class}
\alias{intervalReader-class}
\title{intervalReader}
\description{
This class serves s data structure for storing different configurations
for interval reading.
}
\section{Slots}{

\describe{
\item{\code{numberSeparator}}{string for separating numbers in an interval}

\item{\code{leftClosedBracket}}{left closed bracket string}

\item{\code{rightClosedBracket}}{right closed bracket string}

\item{\code{leftOpenBracket}}{left open bracket string}

\item{\code{rightOpenBracket}}{right open bracket string}

\item{\code{negativeInfinity}}{negative infinity string}

\item{\code{positiveInfinity}}{positive infinity string}

\item{\code{bracketLen}}{how long in chars a bracket is}

\item{\code{decimalSeparator}}{string used as a decimal separator in an interval}
}}

