% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymph}
\alias{lymph}
\title{This is one of three domains provided by the Oncology
Institute that has repeatedly appeared in the machine learning literature. (See also breast-cancer and primary-tumor.)}
\format{A data frame with 130 rows and 19 variables.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Lymphography}
}
\usage{
lymph
}
\description{
This lymphography domain was obtained from the University Medical Centre,
Institute of Oncology, Ljubljana, Yugoslavia. (Restricted access)
}
\keyword{datasets}
