% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{parseItem}
\alias{parseItem}
\title{Function for parsing an item like 'eye_color=blue' or 'height=<166;180)' and
outputting a string explaining the item in natural language like 'eye_color is blue'
or 'height is between 166 to 180'.}
\usage{
parseItem(item, intervalReader)
}
\arguments{
\item{item}{A string containing key and value separated by '=', e.g. 'eye_color=(1;5]'}

\item{intervalReader}{An object of type intervalReader. It allows different types of brackets,
commas and infinity types to be defined and read.}
}
\description{
Function for parsing an item like 'eye_color=blue' or 'height=<166;180)' and
outputting a string explaining the item in natural language like 'eye_color is blue'
or 'height is between 166 to 180'.
}
