% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explainQCBA}
\alias{explainQCBA}
\title{Function for explaning QCBA.}
\usage{
explainQCBA(rulesText, rules, allData, defaultRuleList, intervalReader)
}
\arguments{
\item{rulesText}{Text of the rules.}

\item{rules}{Rule dataframe.}

\item{allData}{Data from which the rules were mined.}

\item{defaultRuleList}{default rules object from qcba model.}

\item{intervalReader}{an intervalReader object used for reading intervals in rules}
}
\value{
dataframe containing explanations for rules
}
\description{
Function for explaning QCBA.
}
