% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanations.R
\name{explainRuleStatistics}
\alias{explainRuleStatistics}
\title{Calculate rules statistics to use in the natural language explanation.}
\usage{
explainRuleStatistics(expl, index, consequentStringTrimmed)
}
\arguments{
\item{index}{Index of the rule we want to calculate statistics for.}

\item{consequentStringTrimmed}{Consequent string trimmed of the surrounding curly brackets.}

\item{allRules}{Dataframe containing all of the rules.}

\item{data}{Data to be used to calculate rule statistics.}
}
\value{
string explaining rule quality.
}
\description{
Calculate rules statistics to use in the natural language explanation.
}
