% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanations.R
\name{getClassExplanationsDataframe}
\alias{getClassExplanationsDataframe}
\title{Function for getting the class explanation dataframe.}
\usage{
getClassExplanationsDataframe(expl, allData)
}
\arguments{
\item{expl}{an explanation object}

\item{allData}{data on which the rule model was trained}
}
\value{
list (where classes are keys) of class explanation dataframes
}
\description{
Function for getting the class explanation dataframe.
}
