% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{getClassExplanationsDataframe}
\alias{getClassExplanationsDataframe}
\title{Function for getting the class explanation dataframe.}
\usage{
getClassExplanationsDataframe(rm, allData, intervalReader)
}
\arguments{
\item{rm}{rule model, either qcba or arc}

\item{allData}{data on which the rule model was trained}

\item{intervalReader}{intervalReader that should be used for reading intervals in
the rules}
}
\value{
list (where classes are keys) of class explanation dataframes
}
\description{
Function for getting the class explanation dataframe.
}
\examples{
  library(arc)

  train <- iris

  rmCBA <- cba(train, classAtt=colnames(train)[length(colnames(train))])

  cbaFiringRuleIDs <- explainPrediction.CBARuleModel(rmCBA, train)
  cbaFiringRules <- as.qcba.rules(rmCBA@rules)[cbaFiringRuleIDs,]

  explanation_dataframe <- getClassExplanationsDataframe(rmCBA, train, createIntervalReader())


}
