% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{getQCBAConflictingRuleText}
\alias{getQCBAConflictingRuleText}
\title{Function for getting information about the next best rule which has different
class from the current rule.}
\usage{
getQCBAConflictingRuleText(rules, ruleIndex, data)
}
\arguments{
\item{rules}{dataframe containing rules information}

\item{ruleIndex}{index of the rule we want to get conflicting text for}

\item{data}{data on which the model was trained}
}
\value{
conflicting rule text
}
\description{
Function for getting information about the next best rule which has different
class from the current rule.
}
